Ajuste no nome do package para projetos productservice e discoveryservice; Inclusão dos métodos de saga para ligar ordens a produtos na hora em que uma ordem fosse lançada; Criação do arqui XStream para evitar erro de forbidden.
