--- DiscoveryServer/src/main/java/com/passdeveloperblog/estore/DiscoveryServer/DiscoveryServerApplication.java
@@ -1,15 +1,0 @@
-package com.passdeveloperblog.estore.DiscoveryServer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@EnableEurekaServer
-@SpringBootApplication
-public class DiscoveryServerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DiscoveryServerApplication.class, args);
-	}
-
-}
