+++ OrdersService-Initial-Assignment/src/main/resources/application.properties
@@ -1,21 +1,42 @@
-server.port=0
-eureka.client.service-url.defaultZone=http://localhost:8761/eureka
 spring.application.name=orders-service
+#Parâmetro para especificar a url de serviço do Eureka Server.
+eureka.client.service-url.defaultZone=http://localhost:8761/eureka/
+
+#Parâmetro que informa ao eureka para usar o endereço IP da máquina ao invés do seu hostname.
+eureka.instance.prefer-ip-address=true
+
+#Parâmetro que permite a aplicação subir em uma porta não usada de forma randomizada.
+server.port=0
+
+#Parâmetro que irá listar cada nova instância da aplicação que subiu no Eureka Server
 eureka.instance.instance-id=${spring.application.name}:${instanceId:${random.value}}
 
-spring.datasource.url=jdbc:h2:file:~/orders-dev;AUTO_SERVER=true
-spring.datasource.username=root
-spring.datasource.password=sergey
+#Parâmetro para especificar o caminho do banco de dados utilizado.
+spring.datasource.url=jdbc:h2:file:~/orders;AUTO_SERVER=true
+
+#Parâmetros para se conectar no banco de dados utilizado
+spring.datasource.username=admin
+spring.datasource.password=admin@123
+
+#Parâmetro de configuração do driver do banco de dados utilizado
 spring.datasource.driver-class-name=org.h2.Driver
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.datasource-platform=org.hibernate.dialect.H2Dialect
+
+#Parâmetro que especifica o schema do banco de dados utilizado
 spring.jpa.hibernate.ddl-auto=update
+
+#Parâmetro que habilita o console do H2 para visualização do banco de dados utilizado
 spring.h2.console.enabled=true
 spring.h2.console.settings.web-allow-others=true
 
+#Parâmetro para incluir a mensagem de erro na resposta HTTP
 server.error.include-message=always
+
+#Parâmetro para configurar quando incluir os erros de binding na resposta HTTP
 server.error.include-binding-errors=always
 
-axon.eventhandling.processors.order-group.mode=subscribing
+#Parâmetro que faz o processing group product-group ser gerenciado por subscribing no event processor
+axon.eventhandling.processors.product-group.mode=subscribing
