+++ ProductsService/src/main/java/com/appsdeveloperblog/estore/ProductsService/command/interceptors/CreateProductCommandInterceptor.java
@@ -1,0 +1,52 @@
+package com.appsdeveloperblog.estore.ProductsService.command.interceptors;
+
+import java.util.List;
+import java.util.function.BiFunction;
+
+import javax.annotation.Nonnull;
+
+import org.axonframework.commandhandling.CommandMessage;
+import org.axonframework.messaging.MessageDispatchInterceptor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Component;
+
+import com.appsdeveloperblog.estore.ProductsService.command.CreateProductCommand;
+import com.appsdeveloperblog.estore.ProductsService.core.data.ProductLookupEntity;
+import com.appsdeveloperblog.estore.ProductsService.core.data.ProductLookupRepository;
+
+@Component
+public class CreateProductCommandInterceptor implements MessageDispatchInterceptor<CommandMessage<?>> {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(CreateProductCommandInterceptor.class);
+    private final ProductLookupRepository productLookupRepository;
+
+    public CreateProductCommandInterceptor(ProductLookupRepository productLookupRepository) {
+        this.productLookupRepository = productLookupRepository;
+    }
+
+    @Override
+    @Nonnull
+    public BiFunction<Integer, CommandMessage<?>, CommandMessage<?>> handle(
+            @Nonnull List<? extends CommandMessage<?>> arg0) {
+        
+        return (index, command) -> {
+            LOGGER.info("Interceptando o comando: " + command.getPayloadType());
+
+            if(CreateProductCommand.class.equals(command.getPayloadType())){
+                CreateProductCommand createProductCommand = (CreateProductCommand) command.getPayload();
+
+                ProductLookupEntity productLookupEntity = productLookupRepository.findByProductIdOrTitle(
+                                                                    createProductCommand.getProductId(), createProductCommand.getTitle());
+                if(productLookupEntity != null) {
+                    throw new IllegalStateException(String.format("Produto com Id %s e Título %s já existente! " , 
+                                                                    createProductCommand.getProductId(), createProductCommand.getTitle()));
+                }
+
+            }
+
+            return command;
+        };
+    }
+
+}
