+++ ProductsService/src/main/java/com/appsdeveloperblog/estore/ProductsService/core/errorhandling/ProductsServiceErrorHandler.java
@@ -1,0 +1,44 @@
+package com.appsdeveloperblog.estore.ProductsService.core.errorhandling;
+
+import java.util.Date;
+
+import org.axonframework.commandhandling.CommandExecutionException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
+
+@ControllerAdvice
+public class ProductsServiceErrorHandler {
+
+    @ExceptionHandler(value = {IllegalStateException.class})
+    public ResponseEntity<Object> handleIllegalStateException(IllegalStateException ex, WebRequest request) {
+        // Log the exception and return a generic error response
+        // This is a placeholder implementation
+
+        ErrorMessage errorMessage = new ErrorMessage(new Date(), ex.getMessage());
+
+        return new ResponseEntity<>(errorMessage, new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    @ExceptionHandler(value = {Exception.class})
+    public ResponseEntity<Object> handleOtherExceptions(Exception ex, WebRequest request) {
+        // Log the exception and return a generic error response
+        // This is a placeholder implementation
+
+        ErrorMessage errorMessage = new ErrorMessage(new Date(), ex.getMessage());
+        return new ResponseEntity<>(errorMessage, new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+    @ExceptionHandler(value = {CommandExecutionException.class})
+    public ResponseEntity<Object> handleCommandExecutionExceptions(CommandExecutionException ex, WebRequest request) {
+        // Log the exception and return a generic error response
+        // This is a placeholder implementation
+
+        ErrorMessage errorMessage = new ErrorMessage(new Date(), ex.getMessage());
+        return new ResponseEntity<>(errorMessage, new HttpHeaders(), HttpStatus.INTERNAL_SERVER_ERROR);
+    }
+
+}
