--- ProductsService/src/main/java/com/passdeveloperblog/estore/ProductsService/command/ProductLookupEventsHandler.java
@@ -1,27 +1,0 @@
-package com.passdeveloperblog.estore.ProductsService.command;
-
-import org.axonframework.config.ProcessingGroup;
-import org.axonframework.eventhandling.EventHandler;
-import org.springframework.stereotype.Component;
-
-import com.passdeveloperblog.estore.ProductsService.core.data.ProductLookupEntity;
-import com.passdeveloperblog.estore.ProductsService.core.data.ProductLookupRepository;
-import com.passdeveloperblog.estore.ProductsService.core.events.ProductCreatedEvent;
-
-@Component
-@ProcessingGroup("product-group")
-public class ProductLookupEventsHandler {
-
-    private final ProductLookupRepository productLookupRepository;
-
-    public ProductLookupEventsHandler(ProductLookupRepository productLookupRepository) {
-        this.productLookupRepository = productLookupRepository;
-    }
-
-    @EventHandler
-    public void on(ProductCreatedEvent event) {
-        ProductLookupEntity productLookupEntity = new ProductLookupEntity(event.getProductId(), event.getTitle());
-
-        productLookupRepository.save(productLookupEntity);
-    }   
-}
