--- ProductsService/src/main/java/com/passdeveloperblog/estore/ProductsService/core/data/ProductLookupRepository.java
@@ -1,9 +1,0 @@
-package com.passdeveloperblog.estore.ProductsService.core.data;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductLookupRepository extends JpaRepository<ProductLookupEntity, String> {
-
-    ProductLookupEntity findByProductIdOrTitle(String productId, String title);
-
-}
