--- ProductsService/src/main/java/com/passdeveloperblog/estore/ProductsService/core/errorhandling/ProductsServiceEventsErrorHandling.java
@@ -1,17 +1,0 @@
-package com.passdeveloperblog.estore.ProductsService.core.errorhandling;
-
-import javax.annotation.Nonnull;
-
-import org.axonframework.eventhandling.EventMessage;
-import org.axonframework.eventhandling.EventMessageHandler;
-import org.axonframework.eventhandling.ListenerInvocationErrorHandler;
-
-public class ProductsServiceEventsErrorHandling implements ListenerInvocationErrorHandler {
-
-    @Override
-    public void onError(@Nonnull Exception exception, @Nonnull EventMessage<?> event, @Nonnull EventMessageHandler eventHandler)
-            throws Exception {
-        throw exception;
-    }
-
-}
