--- ProductsService/src/main/java/com/passdeveloperblog/estore/ProductsService/core/events/ProductCreatedEvent.java
@@ -1,14 +1,0 @@
-package com.passdeveloperblog.estore.ProductsService.core.events;
-
-import java.math.BigDecimal;
-
-import lombok.Data;
-
-@Data
-public class ProductCreatedEvent {
-    
-    private String productId;
-    private String title;
-    private BigDecimal price;
-    private Integer quantity;
-}
