--- ProductsService/src/main/java/com/passdeveloperblog/estore/ProductsService/query/FindProductsQuery.java
@@ -1,5 +1,0 @@
-package com.passdeveloperblog.estore.ProductsService.query;
-
-public class FindProductsQuery {
-
-}
