--- ProductsService/src/main/java/com/passdeveloperblog/estore/ProductsService/query/rest/ProductRestModel.java
@@ -1,15 +1,0 @@
-package com.passdeveloperblog.estore.ProductsService.query.rest;
-
-import java.math.BigDecimal;
-
-import lombok.Data;
-
-@Data
-public class ProductRestModel {
-
-    private String productId;
-    private String title;
-    private BigDecimal price;
-    private Integer quantity;
-
-}
