--- ProductsService/src/main/java/com/passdeveloperblog/estore/ProductsService/query/rest/ProductsQueryController.java
@@ -1,30 +1,0 @@
-package com.passdeveloperblog.estore.ProductsService.query.rest;
-
-import java.util.List;
-
-import org.axonframework.messaging.responsetypes.ResponseTypes;
-import org.axonframework.queryhandling.QueryGateway;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.passdeveloperblog.estore.ProductsService.query.FindProductsQuery;
-
-@RestController
-@RequestMapping("/products")
-public class ProductsQueryController {
-
-    @Autowired
-    QueryGateway queryGateway;
-
-    @GetMapping
-    public List<ProductRestModel> getProducts() {
-        
-        FindProductsQuery findProductsQuery = new FindProductsQuery();
-        List<ProductRestModel> products = queryGateway.query(
-                    findProductsQuery, ResponseTypes.multipleInstancesOf(ProductRestModel.class)).join();
-
-        return products;
-    }
-}
