+++ core/src/main/java/com/appsdeveloperblog/estore/core/config/XStreamConfig.java
@@ -1,0 +1,19 @@
+package com.appsdeveloperblog.estore.core.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import com.thoughtworks.xstream.XStream;
+
+@Configuration
+public class XStreamConfig {
+
+    @Bean
+    XStream xStream() {
+        XStream xStream = new XStream();
+        // Configurações adicionais do XStream, se necessário
+        // Por exemplo, você pode registrar aliases ou deserializar classes específicas
+        xStream.allowTypesByWildcard(new String[] { "com.appsdeveloperblog.**" });
+        return xStream;
+    }
+}
