+++ core/src/main/java/com/appsdeveloperblog/estore/core/events/ProductReservedEvent.java
@@ -1,0 +1,14 @@
+package com.appsdeveloperblog.estore.core.events;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ProductReservedEvent {
+
+    private final String productId;
+    private final Integer quantity;
+    private final String orderId;
+    private final String userId;
+}
